chore: drops process request into an own package.
